requirements.txt
Flask==3.0.0
gunicorn==21.2.0
line-bot-sdk==3.8.0
SQLAlchemy==2.0.25
psycopg[binary]==3.1.18
python-dotenv==1.0.0